Website Update
Renamed files and filled in the comment with name and date. Added correct metadata to the head element. Added a navigation element. Copied text and edited it correctly in the index page. Added a footer to the index. Added metadata to the info page. Copied the navigation from index, and pasted it into the info page. Copied footer from index as well.