Finished FAQ and changed Index
For the FAQ page, created a series of hypertext links between the questions and their answers in the same document. This was done by: marking the 13 questions at the top of the section as an ordered list. Then marked the text of each questions repeated before the answer as an h2 heading and the answer as a paragraph. Added an id attribute to each of the h2 heading, starting with the id faq1, then faq2, down to faq13 for the last question. Also on the FAQ page added a "back to top" link. On the index page, changed the email address to an email link with a subject heading "Entry Form". Changed the word "accommodations" to a hypertext link pointing to the id tag "faq13" in the dr_faq.html file.